Delete superfluous test for definedness of $Htmldir.
Coverage analysis showed that the test suite was never testing the case where
$Htmlroot = '' and not defined $Htmldir.  That's because inside sub
pod2html(), init_globals() is always called before this if block is reached,
and that subroutine always initializes $Htmldir to an empty -- but defined --
string.  Hence, no need to test $Htmldir for definedness at this point, as
it's always already defined.