Add installhtml as of 83c03aa3e02 (May 15 2012).
'installhtml', as it is included in the Perl 5 core distribution, is probably
the most important consumer of Pod::Html.  As it uses the old interface to
Pod::Html::pod2html(), it will eventually have to be revised to use the new OO
interface.  So let's add to this branch a version which we are free to modify.