Change interface to splitpod() to be a single hashref.
This facilitates the complete encapsulation of this subroutine.  Also, this
effort revealed that $htmldir was not used as an argument and that $poddir
could be derived as the dirname of $pod.  Hence, eliminate two arguments.

Rename some variables to reduce confusion.  Move initial declaration of
variables closer to the scope in which they're used.