split_on_head(): Change interface to be a single hashref.
This facilitates the complete encapsulation of this subroutine.  Also, this
effort revealed that certain arguments were not needed.  Specifically,
splitdirs and ignoredirs did not need to be declared above this sub's
invocation because they are only first assigned to within this function.

Begin to rework split_on_item() along similar lines.