Abortive attempt at standalone
Replacing all the *.pm and *.t files and replacing t/pod2html-lib.pl
with t/lib/Testing.pm is not working.  The BEGIN blocks in the test
files, plus (I think) setup_testing-directory(), means that @INC is not
getting populated correctly.

My motivation for creating this branch was to be able to do coverage
analysis on this.  Will have to defer attempting that for now.  (And
even when/if I get back to it, I face the problem of inability to use
Devel::Cover for distros that ship with core.  I'll have to recall how I
tried to do that back in 2012.)