Compare hashes, not just list of keys
Test failure was probably due to impact of hash randomization
implemented after this test was originally written.  We can set an
expectation for a sorted list of hash keys or for the full hash itself,
but no longer for an unsorted list of hash keys.

At this point we have failing tests in other files, but we're not going
to worry about them now.